Add FPS limiter. This also splits out screen functions into its own c file.
